api/requirements.txt
Flask
google-cloud-firestore
pandas
numpy
alpaca-py
requests
python-dotenv